ability to set active driver
In cases where multiple database drivers are registered, it is good to be able to set the correct one.

This solution is bad, however, as it uses global mutable state and is not safe for concurrent use. Better solution would be to wrap each query/statement execution with its own specified driver name. Better yet would be to detect the driver from the passed in Querier/Executor object.