Test suite (#1)
* test module

* ability to set active driver

In cases where multiple database drivers are registered, it is good to be able to set the correct one.

This solution is bad, however, as it uses global mutable state and is not safe for concurrent use. Better solution would be to wrap each query/statement execution with its own specified driver name. Better yet would be to detect the driver from the passed in Querier/Executor object.

* basic module for running integration tests

Testing against actual databases is preferable. This is a basic setup using docker-compose to bring up databases and then run tests.
Using the 'testcontainers-go' library to bring up docker-compose in code to enable debugging.

* sqlite3 integration tests

* github workflow

* sqlite3 positionals

TODO: this is wrong. This assumes that sqlite3 only supports '?' as the positional parameter. Need to support multiple.

* added register for driver in tests

Tests fail otherwise. A driver needs to be registered for stuff to work.

* '?' as the positional parameter for sqlite3

In the interim until support for all is added.